Move VELOX_FORMAT_CONFIG and registerConfigProperty to common place
